requirements.txt
Django==3.0.5
python-dateutil==2.8.1
pandas==1.0.3
